Centralise interfaces in `domain/interfaces`
