+++ src/domain/Command.ts
@@ -1,18 +1,4 @@
-export type IRejectFunction = (reason: string) => void;
-
-export interface IDomainCommand {
-  // Name of the command
-  name: string;
-
-  // Parameters associated with this command
-  data?: any;
-
-  // User initiating the command
-  user?: object;
-
-  // Hook to reject the command if it isn't valid
-  reject: (reason: string) => void;
-}
+import { IDomainCommand, IRejectFunction } from './interfaces';
 
 export function createCommand(
   name: string,
