+++ src/domain/Entity.ts
@@ -1,15 +1,7 @@
+import { IVersionedEntity } from './interfaces';
+
 interface IParams<T> {
   state: T;
-  version: number;
-}
-
-export interface IVersionedEntity<T> {
-  // Current state of the entity
-  state: T;
-
-  update?: (state: T) => IVersionedEntity<T>;
-
-  // Current version of the entity
   version: number;
 }
