+++ src/domain/Event.ts
@@ -1,10 +1,4 @@
-export interface IDomainEvent {
-  // Name of this event
-  name: string;
-
-  // Data associated with this event
-  data?: any;
-}
+import { IDomainEvent } from './interfaces';
 
 export function createEvent(name: string, data?: object): IDomainEvent {
   return { name, data };
