--- src/domain/Service.ts
@@ -1,7 +1,0 @@
-export interface IDomainService {
-  [s: string]: (...args: any) => Promise<any>;
-}
-
-export interface IServiceRegistry {
-  [s: string]: IDomainService;
-}
